Code Clean Up + prerequisites for input hints
