+++ src/Mod/TechDraw/Gui/CommandExtensionPack.cpp
@@ -72,218 +72,19 @@
 #include "ViewProviderBalloon.h"
 #include "ViewProviderDimension.h"
 #include "ViewProviderPage.h"
+#include "CommandHelpers.h"
 #include "TechDrawHandler.h"
+#include "TechDrawCircleCenterlinesHandler.h"
 
 
 using namespace TechDrawGui;
 using namespace TechDraw;
 using DU = DrawUtil;
 
-
-namespace TechDrawGui
-{
-//TechDraw::LineFormat activeAttributes; // container holding global line attributes
-
-//internal helper functions
-TechDraw::LineFormat& _getActiveLineAttributes();
-Base::Vector3d _circleCenter(Base::Vector3d p1, Base::Vector3d p2, Base::Vector3d p3);
-void _createThreadCircle(const std::string Name, TechDraw::DrawViewPart* objFeat, double factor);
-void _createThreadLines(const std::vector<std::string>& SubNames, TechDraw::DrawViewPart* objFeat,
-                        double factor, bool endLine);
-void _setLineAttributes(TechDraw::CosmeticEdge* cosEdge);
-void _setLineAttributes(TechDraw::CenterLine* cosEdge);
-void _setLineAttributes(TechDraw::CosmeticEdge* cosEdge, int style, float weight, Base::Color color);
-void _setLineAttributes(TechDraw::CenterLine* cosEdge, int style, float weight, Base::Color color);
-double _getAngle(Base::Vector3d center, Base::Vector3d point);
-std::vector<Base::Vector3d> _getVertexPoints(const std::vector<std::string>& SubNames,
-                                             TechDraw::DrawViewPart* objFeat);
-bool _checkSel(Gui::Command* cmd, std::vector<Gui::SelectionObject>& selection,
-               TechDraw::DrawViewPart*& objFeat, const std::string& message);
-std::string _createBalloon(Gui::Command* cmd, TechDraw::DrawViewPart* objFeat);
-
 //===========================================================================
 // TechDraw_ExtensionHoleCircle
 //===========================================================================
 
-void execHoleCircle(std::vector<std::string> SubNames, TechDraw::DrawViewPart* objFeat, 
-    std::vector<std::string>& lines)
-{
-    //create centerlines of a hole/bolt circle
-
-    std::vector<TechDraw::CirclePtr> Circles;
-    for (const std::string& Name : SubNames) {
-        int GeoId = TechDraw::DrawUtil::getIndexFromName(Name);
-        std::string GeoType = TechDraw::DrawUtil::getGeomTypeFromName(Name);
-        TechDraw::BaseGeomPtr geom = objFeat->getGeomByIndex(GeoId);
-        if (GeoType == "Edge") {
-            if (geom->getGeomType() == GeomType::CIRCLE || geom->getGeomType() == GeomType::ARCOFCIRCLE) {
-                TechDraw::CirclePtr cgen = std::static_pointer_cast<TechDraw::Circle>(geom);
-                Circles.push_back(cgen);
-            }
-        }
-    }
-
-    if (Circles.size() <= 2) {
-        return;
-    }
-
-    // make the bolt hole circle from 3 scaled and rotated points
-    Base::Vector3d bigCenter =
-        _circleCenter(Circles[0]->center, Circles[1]->center, Circles[2]->center);
-    double bigRadius = (Circles[0]->center - bigCenter).Length();
-    // now convert the center & radius to canonical form
-    bigCenter = CosmeticVertex::makeCanonicalPointInverted(objFeat, bigCenter);
-    bigRadius = bigRadius / objFeat->getScale();
-    TechDraw::BaseGeomPtr bigCircle =
-        std::make_shared<TechDraw::Circle>(bigCenter, bigRadius);
-    std::string bigCircleTag = objFeat->addCosmeticEdge(bigCircle);
-    lines.push_back(bigCircleTag);
-    TechDraw::CosmeticEdge* ceCircle = objFeat->getCosmeticEdge(bigCircleTag);
-    _setLineAttributes(ceCircle);
-
-    // make the center lines for the individual bolt holes
-    constexpr double ExtendFactor{1.1};
-    for (const TechDraw::CirclePtr& oneCircle : Circles) {
-        // convert the center to canonical form
-        Base::Vector3d oneCircleCenter = CosmeticVertex::makeCanonicalPointInverted(objFeat, oneCircle->center);
-        // oneCircle->radius is scaled.
-        double oneRadius = oneCircle->radius / objFeat->getScale();
-        // what is magic number 2 (now ExtendFactor)?  just a fudge factor to extend the line beyond the bolt
-        // hole circle?  should it be a function of hole diameter? maybe 110% of oneRadius?
-        Base::Vector3d delta = (oneCircleCenter - bigCenter).Normalize() * (oneRadius * ExtendFactor);
-        Base::Vector3d startPt = oneCircleCenter + delta;
-        Base::Vector3d endPt = oneCircleCenter - delta;
-        std::string oneLineTag = objFeat->addCosmeticEdge(startPt, endPt);
-        TechDraw::CosmeticEdge* ceLine = objFeat->getCosmeticEdge(oneLineTag);
-        lines.push_back(oneLineTag);
-        _setLineAttributes(ceLine);
-    }
-    objFeat->refreshCEGeoms();
-    objFeat->requestPaint();
-}
-}// namespace TechDrawGui
-
-
-class TDBoltCenterlinesHandler : public TechDrawHandler {
-    private:
-        std::vector<std::string> currentLines;
-        TechDraw::DrawViewPart* currentObjFeat{nullptr};
-        std::vector<std::string> previousSubnames;
-    public:
-        void activated() override
-        {
-            if (viewPage) {
-                QPoint hotspot(15, 15);
-                QPixmap pixmap = viewPage->prepareCursorPixmap("TechDraw_ExtensionHoleCircle_Pointer.svg", hotspot);
-                QCursor cursor(pixmap, hotspot.x(), hotspot.y());
-                if (auto vp = viewPage->viewport()) {
-                    vp->setCursor(cursor);
-                }
-            }
-
-            Gui::Selection().setSelectionStyle(Gui::SelectionSingleton::SelectionStyle::GreedySelection);
-        }
-
-        void deactivated() override
-        {
-            Gui::Selection().setSelectionStyle(Gui::SelectionSingleton::SelectionStyle::NormalSelection);
-            if (viewPage) {
-                if (auto vp = viewPage->viewport()) {
-                    vp->unsetCursor();
-                }
-            }
-            currentLines.clear();
-            currentObjFeat = nullptr;
-        }
-
-        void mouseMoveEvent(QMouseEvent* event) override
-        {
-            Q_UNUSED(event);
-        }
-
-        void mouseReleaseEvent(QMouseEvent* event) override
-        {
-            std::vector<Gui::SelectionObject> selection = Gui::Selection().getSelectionEx();
-
-            TechDraw::DrawViewPart* objFeat{nullptr};
-            std::vector<std::string> subNames;
-
-            if (!selection.empty()) {
-                objFeat = dynamic_cast<TechDraw::DrawViewPart*>(selection[0].getObject());
-                subNames = selection[0].getSubNames();
-
-                // If the user clicks on cosmetic edges we stop the selection
-                // So we dont attemt to reference it when we remove the former lines
-                if (!currentLines.empty()) {
-                    for (const std::string& name : subNames) {
-                        TechDraw::CosmeticEdge* cosmeticEdge = objFeat->getCosmeticEdgeBySelection(name);
-                        if (cosmeticEdge) {
-                            previousSubnames.clear();
-                            currentLines.clear();
-                            currentObjFeat = nullptr;
-                            Gui::Selection().clearSelection();
-                            return;
-                        }
-                    }
-                }
-
-            } else {
-                return;
-            }
-
-            // If the selection does not change then they have finished selecting circles
-            // It is possible to do this by just looking at the size of the selection since it is greedy selection
-            if (subNames.size() == previousSubnames.size()) {
-                previousSubnames.clear();
-                Gui::Selection().clearSelection();
-                currentLines.clear();
-                currentObjFeat = nullptr;
-                return;
-            }
-
-            previousSubnames = subNames;
-
-            // The redrawing of the lines clears the selection somehow
-            // But it works if it just gets blocked before we redraw
-            auto* mdi = dynamic_cast<MDIViewPage*>(Gui::getMainWindow()->activeWindow());
-            if (mdi) {
-                mdi->blockSceneSelection(true);
-            }
-            
-            if (!currentLines.empty() && currentObjFeat) {
-                currentObjFeat->removeCosmeticEdge(currentLines);
-                currentObjFeat->refreshCEGeoms();
-                currentObjFeat->requestPaint();
-                currentLines.clear();
-                currentObjFeat = nullptr;
-            }
-    
-            if (objFeat) {
-                currentObjFeat = objFeat;
-
-                execHoleCircle(subNames, objFeat, currentLines);
-            }
-
-            if (mdi) {
-                if (objFeat) {
-                    mdi->selectQGIView(objFeat, true, subNames);
-                }
-                mdi->blockSceneSelection(false);
-            }
-
-            Q_UNUSED(event);
-        }
-
-        void mousePressEvent(QMouseEvent* event) override
-        {
-            Q_UNUSED(event);
-        }
-
-        void keyPressEvent(QKeyEvent* event) override
-        {
-            Q_UNUSED(event);
-        }
-};
 
 DEF_STD_CMD_A(CmdTechDrawExtensionHoleCircle)
 
@@ -318,18 +119,12 @@ void CmdTechDrawExtensionHoleCircle::activated(int iMsg)
         return;
     }
 
+    TechDrawBoltCenterlinesHandler* handler = new TechDrawBoltCenterlinesHandler();
+
     // Post selection handler
-    viewPage->activateHandler(new TDBoltCenterlinesHandler());
+    viewPage->activateHandler(handler);
 
-    std::vector<Gui::SelectionObject> selection = Gui::Selection().getSelectionEx();
-    
-    if (!selection.empty()) {
-        TechDraw::DrawViewPart* objFeat = dynamic_cast<TechDraw::DrawViewPart*>(selection[0].getObject());
-        std::vector<std::string> SubNames = selection[0].getSubNames();
-
-        std::vector<std::string> currentLines;
-        execHoleCircle(SubNames, objFeat, currentLines);
-    }
+    handler->addPreselected();
 }
 
 bool CmdTechDrawExtensionHoleCircle::isActive()
@@ -342,164 +137,6 @@ bool CmdTechDrawExtensionHoleCircle::isActive()
 //===========================================================================
 // TechDraw_ExtensionCircleCenterLines
 //===========================================================================
-
-void execCircleCenterLines(std::vector<std::string> SubNames, TechDraw::DrawViewPart* objFeat,
-                           std::vector<std::string>& lines)
-{
-    // create circle centerlines
-    for (const std::string& Name : SubNames) {
-        int GeoId = TechDraw::DrawUtil::getIndexFromName(Name);
-        TechDraw::BaseGeomPtr geom = objFeat->getGeomByIndex(GeoId);
-        std::string GeoType = TechDraw::DrawUtil::getGeomTypeFromName(Name);
-        if (GeoType == "Edge") {
-            if (!geom) {
-                continue;
-            }
-            if (geom->getGeomType() == GeomType::CIRCLE || geom->getGeomType() == GeomType::ARCOFCIRCLE) {
-                TechDraw::CirclePtr cgen = std::static_pointer_cast<TechDraw::Circle>(geom);
-                // cgen->center is a scaled, rotated and inverted point
-                Base::Vector3d center = CosmeticVertex::makeCanonicalPointInverted(objFeat, cgen->center);
-                double radius = cgen->radius / objFeat->getScale();
-                // right, left, top, bottom are formed from a canonical point (center)
-                // so they do not need to be changed to canonical form.
-                constexpr double lineOutsideCircle{2.0};
-                Base::Vector3d right(center.x + radius + lineOutsideCircle, center.y, 0.0);
-                Base::Vector3d top(center.x, center.y + radius + lineOutsideCircle, 0.0);
-                Base::Vector3d left(center.x - radius - lineOutsideCircle, center.y, 0.0);
-                Base::Vector3d bottom(center.x, center.y - radius - lineOutsideCircle, 0.0);
-                std::string line1tag = objFeat->addCosmeticEdge(right, left);
-                std::string line2tag = objFeat->addCosmeticEdge(top, bottom);
-                lines.push_back(line1tag);
-                lines.push_back(line2tag);
-                TechDraw::CosmeticEdge* horiz = objFeat->getCosmeticEdge(line1tag);
-                _setLineAttributes(horiz);
-                TechDraw::CosmeticEdge* vert = objFeat->getCosmeticEdge(line2tag);
-                _setLineAttributes(vert);
-                // horiz & vert are centerlines, so they should use the default centerline
-                // number and not the number from line attributes
-                horiz->m_format.setLineNumber(Preferences::CenterLineStyle());
-                vert->m_format.setLineNumber(Preferences::CenterLineStyle());
-            }
-        }
-    }
-
-    objFeat->refreshCEGeoms();
-    objFeat->requestPaint();
-}
-
-class TDCircleCenterlinesHandler : public TechDrawHandler {
-    private:
-        std::vector<std::string> previewLines;
-        TechDraw::DrawViewPart* previewObjFeat{nullptr};
-    public:
-        void activated() override
-        {
-            if (viewPage) {
-                QPoint hotspot(15, 15);
-                QPixmap pixmap = viewPage->prepareCursorPixmap("TechDraw_ExtensionCircleCenterLines_Pointer.svg", hotspot);
-                QCursor cursor(pixmap, hotspot.x(), hotspot.y());
-                if (auto vp = viewPage->viewport()) {
-                    vp->setCursor(cursor);
-                }
-            }
-        }
-
-        void deactivated() override
-        {
-            if (viewPage) {
-                if (auto vp = viewPage->viewport()) {
-                    vp->unsetCursor();
-                }
-            }
-            if (!previewLines.empty() && previewObjFeat) {
-                previewObjFeat->removeCosmeticEdge(previewLines);
-                previewObjFeat->refreshCEGeoms();
-                previewObjFeat->requestPaint();
-                previewLines.clear();
-                previewObjFeat = nullptr;
-            }
-        }
-
-        void mouseMoveEvent(QMouseEvent* event) override
-        {
-            if (!viewPage) {
-                return;
-            }
-            auto items = viewPage->scene()->items(viewPage->mapToScene(event->pos()));
-            
-            std::vector<std::string> SubNames;
-            TechDraw::DrawViewPart* objFeat{nullptr};
-
-            for (QGraphicsItem* item : items) {
-                QGIEdge* edge = dynamic_cast<QGIEdge*>(item);
-                if (edge) {
-                    SubNames.push_back("Edge" + std::to_string(edge->getProjIndex()));
-                    auto* parent = edge->parentItem();
-                    QGIView* view = dynamic_cast<QGIView*>(parent);
-                    if (view) {
-                        objFeat = dynamic_cast<TechDraw::DrawViewPart*>(view->getViewObject());
-                    }
-
-                    int GeoId = TechDraw::DrawUtil::getIndexFromName("Edge" + std::to_string(edge->getProjIndex()));
-                    TechDraw::BaseGeomPtr geom = objFeat->getGeomByIndex(GeoId);
-
-                    if (geom->getGeomType() != GeomType::CIRCLE && geom->getGeomType() != GeomType::ARCOFCIRCLE) {
-                        SubNames.clear();
-                        objFeat = nullptr;
-                        continue;
-                    }
-                    break;
-                }
-            }
-
-            if (SubNames.empty() || !objFeat) {
-                if (!previewLines.empty()) {
-                    previewObjFeat->removeCosmeticEdge(previewLines);
-                    previewObjFeat->refreshCEGeoms();
-                    previewObjFeat->requestPaint();
-                    previewLines.clear();
-                    previewObjFeat = nullptr;
-                }
-            }
-            else {
-                execCircleCenterLines(SubNames, objFeat, previewLines);
-                previewObjFeat = objFeat;
-            }
-        }
-
-        void mouseReleaseEvent(QMouseEvent* event) override
-        {
-            std::vector<Gui::SelectionObject> selection = Gui::Selection().getSelectionEx();
-    
-            if (!selection.empty()) {
-                TechDraw::DrawViewPart* objFeat = dynamic_cast<TechDraw::DrawViewPart*>(selection[0].getObject());
-                std::vector<std::string> SubNames = selection[0].getSubNames();
-                std::vector<std::string> lines;
-
-                int GeoId = TechDraw::DrawUtil::getIndexFromName(SubNames[0]);
-                TechDraw::BaseGeomPtr geom = objFeat->getGeomByIndex(GeoId);
-
-                if (geom->getGeomType() != GeomType::CIRCLE && geom->getGeomType() != GeomType::ARCOFCIRCLE) {
-                    return;
-                }
-
-                execCircleCenterLines(SubNames, objFeat, lines);
-            }
-
-            Gui::Selection().clearSelection();
-        }
-
-        void mousePressEvent(QMouseEvent* event) override
-        {
-            Q_UNUSED(event);
-        }
-
-        void keyPressEvent(QKeyEvent* event) override
-        {
-            Q_UNUSED(event);
-        }
-
-};
 
 DEF_STD_CMD_A(CmdTechDrawExtensionCircleCenterLines)
 
@@ -534,19 +171,12 @@ void CmdTechDrawExtensionCircleCenterLines::activated(int iMsg)
         return;
     }
 
-    // Post selection handler
-    viewPage->activateHandler(new TDCircleCenterlinesHandler());
+    TechDrawCircleCenterlinesHandler* handler = new TechDrawCircleCenterlinesHandler();
 
-    // Preselection handling
-    std::vector<Gui::SelectionObject> selection = getSelection().getSelectionEx();
-    
-    if (!selection.empty()) {
-        TechDraw::DrawViewPart* objFeat = dynamic_cast<TechDraw::DrawViewPart*>(selection[0].getObject());
-        std::vector<std::string> SubNames = selection[0].getSubNames();
-        std::vector<std::string> lines;
-        execCircleCenterLines(SubNames, objFeat, lines);
-    }
-    
+    // Post selection handler
+    viewPage->activateHandler(handler);
+
+    handler->addPreselected();
 }
 
 bool CmdTechDrawExtensionCircleCenterLines::isActive()
@@ -2379,224 +2009,6 @@ bool CmdTechDrawExtensionArcLengthAnnotation::isActive()
     bool haveView = DrawGuiUtil::needView(this);
     return (havePage && haveView);
 }
-
-//===========================================================================
-// internal helper routines
-//===========================================================================
-namespace TechDrawGui
-{
-
-LineFormat& _getActiveLineAttributes()
-{
-    return LineFormat::getCurrentLineFormat();
-}
-
-std::string _createBalloon(Gui::Command* cmd, TechDraw::DrawViewPart* objFeat)
-// create a new balloon, return its name as string
-{
-    std::string featName;
-    TechDraw::DrawPage* page = objFeat->findParentPage();
-    Gui::Document* guiDoc = Gui::Application::Instance->getDocument(page->getDocument());
-    auto pageVP = freecad_cast<ViewProviderPage*>(guiDoc->getViewProvider(page));
-    if (pageVP) {
-        QGSPage* scenePage = pageVP->getQGSPage();
-        featName = scenePage->getDrawPage()->getDocument()->getUniqueObjectName("Balloon");
-        std::string pageName = scenePage->getDrawPage()->getNameInDocument();
-        cmd->doCommand(cmd->Doc,
-                       "App.activeDocument().addObject('TechDraw::DrawViewBalloon', '%s')",
-                       featName.c_str());
-        cmd->doCommand(cmd->Doc, "App.activeDocument().%s.SourceView = (App.activeDocument().%s)",
-                       featName.c_str(), objFeat->getNameInDocument());
-
-        cmd->doCommand(cmd->Doc, "App.activeDocument().%s.addView(App.activeDocument().%s)",
-                       pageName.c_str(), featName.c_str());
-    }
-    return featName;
-}
-
-bool _checkSel(Gui::Command* cmd, std::vector<Gui::SelectionObject>& selection,
-               TechDraw::DrawViewPart*& objFeat, const std::string& message)
-{
-    // check selection of getSelectionEx() and selection[0].getObject()
-    selection = cmd->getSelection().getSelectionEx();
-    if (selection.empty()) {
-        // message is translated in caller
-        QMessageBox::warning(Gui::getMainWindow(), QString::fromUtf8(message.c_str()),
-                             QObject::tr("Selection is empty"));
-        return false;
-    }
-
-    objFeat = dynamic_cast<TechDraw::DrawViewPart*>(selection[0].getObject());
-    if (!objFeat) {
-        QMessageBox::warning(Gui::getMainWindow(), QString::fromUtf8(message.c_str()),
-                             QObject::tr("No object selected"));
-        return false;
-    }
-
-    return true;
-}
-
-//! return the vertices in the selection as [Base::Vector3d]
-std::vector<Base::Vector3d> _getVertexPoints(const std::vector<std::string>& SubNames,
-                                             TechDraw::DrawViewPart* objFeat)
-{
-    std::vector<Base::Vector3d> vertexPoints;
-    for (const std::string& Name : SubNames) {
-        std::string GeoType = TechDraw::DrawUtil::getGeomTypeFromName(Name);
-        if (GeoType == "Vertex") {
-            int GeoId = TechDraw::DrawUtil::getIndexFromName(Name);
-            TechDraw::VertexPtr vert = objFeat->getProjVertexByIndex(GeoId);
-            vertexPoints.push_back(vert->point());
-        }
-    }
-    return vertexPoints;
-}
-
-//! get angle between x-axis and the vector from center to point.
-//! result is [0, 360]
-double _getAngle(Base::Vector3d center, Base::Vector3d point)
-{
-    constexpr double DegreesHalfCircle{180.0};
-    Base::Vector3d vecCP = point - center;
-    double angle = DU::angleWithX(vecCP) * DegreesHalfCircle / std::numbers::pi;
-    return angle;
-}
-
-Base::Vector3d _circleCenter(Base::Vector3d p1, Base::Vector3d p2, Base::Vector3d p3)
-{
-    Base::Vector2d v1(p1.x, p1.y);
-    Base::Vector2d v2(p2.x, p2.y);
-    Base::Vector2d v3(p3.x, p3.y);
-    Base::Vector2d center = Part::Geom2dCircle::getCircleCenter(v1, v2, v3);
-    return Base::Vector3d(center.x, center.y, 0.0);
-}
-
-void _createThreadCircle(const std::string Name, TechDraw::DrawViewPart* objFeat, double factor)
-{
-    constexpr double ArcStartDegree{15.0};
-    constexpr double ArcEndDegree{285.0};
-    // create the 3/4 arc symbolizing a thread from top seen
-    double scale = objFeat->getScale();
-    int GeoId = TechDraw::DrawUtil::getIndexFromName(Name);
-    TechDraw::BaseGeomPtr geom = objFeat->getGeomByIndex(GeoId);
-    std::string GeoType = TechDraw::DrawUtil::getGeomTypeFromName(Name);
-
-    if (GeoType == "Edge" && geom->getGeomType() == GeomType::CIRCLE) {
-        TechDraw::CirclePtr cgen = std::static_pointer_cast<TechDraw::Circle>(geom);
-        // center is rotated and scaled
-        Base::Vector3d center = CosmeticVertex::makeCanonicalPointInverted(objFeat, cgen->center);
-        // radius is scaled
-        float radius = cgen->radius * factor / scale;
-        TechDraw::BaseGeomPtr threadArc =
-            std::make_shared<TechDraw::AOC>(center, radius, ArcStartDegree, ArcEndDegree);
-        std::string arcTag = objFeat->addCosmeticEdge(threadArc);
-        TechDraw::CosmeticEdge* arc = objFeat->getCosmeticEdge(arcTag);
-        int solidStyle = 1; // Qt::SolidLine
-        float thinWeight = (float)TechDraw::DrawUtil::getDefaultLineWeight("Thin");
-        Base::Color threadColor = _getActiveLineAttributes().getColor(); 
-        _setLineAttributes(arc, solidStyle, thinWeight, threadColor);
-    } else {
-        QMessageBox::warning(Gui::getMainWindow(), QObject::tr("TechDraw create thread circle"),
-                             QObject::tr("Can not make thread circle for %1")
-                                 .arg(QString::fromStdString(GeometryUtils::getGeomTypeName(geom->getGeomType()))));
-    }
-}
-
-void _createThreadLines(const std::vector<std::string>& SubNames, TechDraw::DrawViewPart* objFeat,
-                        double factor, bool endLine)
-{
-    // create symbolizing lines of a thread from the side seen
-    std::string GeoType0 = TechDraw::DrawUtil::getGeomTypeFromName(SubNames[0]);
-    std::string GeoType1 = TechDraw::DrawUtil::getGeomTypeFromName(SubNames[1]);
-    if ((GeoType0 == "Edge") && (GeoType1 == "Edge")) {
-        int GeoId0 = TechDraw::DrawUtil::getIndexFromName(SubNames[0]);
-        int GeoId1 = TechDraw::DrawUtil::getIndexFromName(SubNames[1]);
-        TechDraw::BaseGeomPtr geom0 = objFeat->getGeomByIndex(GeoId0);
-        TechDraw::BaseGeomPtr geom1 = objFeat->getGeomByIndex(GeoId1);
-        if (geom0->getGeomType() != GeomType::GENERIC || geom1->getGeomType() != GeomType::GENERIC) {
-            QMessageBox::warning(Gui::getMainWindow(), QObject::tr("TechDraw thread hole side"),
-                                 QObject::tr("Select 2 straight lines"));
-            return;
-        }
-
-        TechDraw::GenericPtr line0 = std::static_pointer_cast<TechDraw::Generic>(geom0);
-        TechDraw::GenericPtr line1 = std::static_pointer_cast<TechDraw::Generic>(geom1);
-        // start and end points are scaled,rotated and inverted (CSRIx).
-        // convert start and end to unscaled, unrotated.
-        Base::Vector3d start0 = CosmeticVertex::makeCanonicalPointInverted(objFeat, line0->getStartPoint());
-        Base::Vector3d start1 = CosmeticVertex::makeCanonicalPointInverted(objFeat, line1->getStartPoint());
-        Base::Vector3d end0 = CosmeticVertex::makeCanonicalPointInverted(objFeat, line0->getEndPoint());
-        Base::Vector3d end1 = CosmeticVertex::makeCanonicalPointInverted(objFeat, line1->getEndPoint());
-        if (DrawUtil::circulation(start0, end0, start1)
-            != DrawUtil::circulation(end0, end1, start1)) {
-            Base::Vector3d help1 = start1;
-            Base::Vector3d help2 = end1;
-            start1 = help2;
-            end1 = help1;
-        }
-        float kernelDiam = (start1 - start0).Length();
-        float kernelFactor = (kernelDiam * factor - kernelDiam) / 2;
-        Base::Vector3d delta = (start1 - start0).Normalize() * kernelFactor;
-        std::string line0Tag =
-            objFeat->addCosmeticEdge(start0 - delta, end0 - delta);
-        std::string line1Tag =
-            objFeat->addCosmeticEdge(start1 + delta, end1 + delta);
-        TechDraw::CosmeticEdge* cosTag0 = objFeat->getCosmeticEdge(line0Tag);
-        TechDraw::CosmeticEdge* cosTag1 = objFeat->getCosmeticEdge(line1Tag);
-        int solidStyle = Qt::SolidLine;
-        float thinWeight = (float)TechDraw::DrawUtil::getDefaultLineWeight("Thin");
-        Base::Color threadColor = _getActiveLineAttributes().getColor();
-        _setLineAttributes(cosTag0, solidStyle, thinWeight, threadColor);
-        _setLineAttributes(cosTag1, solidStyle, thinWeight, threadColor);
-        if (endLine) {
-            float graphicWeight = (float)TechDraw::DrawUtil::getDefaultLineWeight("Graphic");
-            std::string line3Tag =
-                objFeat->addCosmeticEdge(end0 - delta, end1 + delta);
-            TechDraw::CosmeticEdge* cosTag3 = objFeat->getCosmeticEdge(line3Tag);
-            _setLineAttributes(cosTag3, solidStyle, graphicWeight, threadColor);
-        }
-    }
-}
-
-void _setLineAttributes(TechDraw::CosmeticEdge* cosEdge)
-{
-    // set line attributes of a cosmetic edge
-    cosEdge->m_format.setStyle(_getActiveLineAttributes().getStyle());
-    cosEdge->m_format.setWidth(_getActiveLineAttributes().getWidth());
-    cosEdge->m_format.setColor(_getActiveLineAttributes().getColor());
-    cosEdge->m_format.setVisible(_getActiveLineAttributes().getVisible());
-    cosEdge->m_format.setLineNumber(_getActiveLineAttributes().getLineNumber());
-}
-
-void _setLineAttributes(TechDraw::CenterLine* cosEdge)
-{
-    // set line attributes of a cosmetic edge
-    cosEdge->m_format.setStyle(_getActiveLineAttributes().getStyle());
-    cosEdge->m_format.setWidth(_getActiveLineAttributes().getWidth());
-    cosEdge->m_format.setColor(_getActiveLineAttributes().getColor());
-    cosEdge->m_format.setVisible(_getActiveLineAttributes().getVisible());
-    cosEdge->m_format.setLineNumber(_getActiveLineAttributes().getLineNumber());
-}
-
-void _setLineAttributes(TechDraw::CosmeticEdge* cosEdge, int style, float weight, Base::Color color)
-{
-    // set line attributes of a cosmetic edge
-    cosEdge->m_format.setStyle(style);
-    cosEdge->m_format.setWidth(weight);
-    cosEdge->m_format.setColor(color);
-    cosEdge->m_format.setVisible(_getActiveLineAttributes().getVisible());
-    cosEdge->m_format.setLineNumber(style);
-}
-
-void _setLineAttributes(TechDraw::CenterLine* cosEdge, int style, float weight, Base::Color color)
-{
-    // set line attributes of a centerline
-    cosEdge->m_format.setStyle(style);
-    cosEdge->m_format.setWidth(weight);
-    cosEdge->m_format.setColor(color);
-    cosEdge->m_format.setVisible(_getActiveLineAttributes().getVisible());
-    cosEdge->m_format.setLineNumber(style);}
-}// namespace TechDrawGui
 
 //------------------------------------------------------------------------------
 void CreateTechDrawCommandsExtensions()
