+++ src/Mod/TechDraw/Gui/QGIEdge.cpp
@@ -23,6 +23,8 @@
 
 # include <QPainterPath>
 # include <QPainterPathStroker>
+# include <QStyleOptionGraphicsItem>
+
 # include <cmath>
