+++ src/Mod/TechDraw/Gui/TechDrawHandler.cpp
@@ -37,6 +37,7 @@
 #include <Gui/MainWindow.h>
 #include <Gui/View3DInventor.h>
 #include <Gui/View3DInventorViewer.h>
+#include <Gui/InputHint.h>
 
 #include "MDIViewPage.h"
 #include "QGVPage.h"
@@ -66,6 +67,12 @@ void TechDrawHandler::activate(QGVPage* vp)
     if (!Gui::ToolHandler::activate()) {
         viewPage->deactivateHandler();
     }
+}
+
+
+std::list<Gui::InputHint> TechDrawHandler::getToolHints() const
+{
+    return {};
 }
 
 void TechDrawHandler::deactivate()
