+++ src/Mod/TechDraw/Gui/TechDrawHandler.h
@@ -22,8 +22,13 @@
 
 #pragma once
 
+
+#include <QKeyEvent>
+#include <QMouseEvent>
+
 #include <Gui/ToolHandler.h>
 
+#include <Mod/TechDraw/App/DrawPage.h>
 #include <Mod/TechDraw/TechDrawGlobal.h>
 
 namespace TechDrawGui
@@ -41,6 +46,8 @@ public:
     void deactivate() override;
 
     void quit() override;
+
+    std::list<Gui::InputHint> getToolHints() const override;
 
     virtual void mouseMoveEvent(QMouseEvent* event) = 0;
     virtual void mousePressEvent(QMouseEvent* event) { Q_UNUSED(event) };
